requirements.txt
Flask>=2.0
gradio_client>=0.10.0
python-dotenv>=1.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.28.0
pytest>=7.0.0
pytest-cov>=4.1.0
